pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-dashboard"
version = "1.0.0"
description = "Beautiful dashboard-style HTML reports for pytest with charts, error analysis, and visual insights"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Niresh Shanmugam", email = "niresh.shanmugam@gmail.com" }
]
keywords = ["pytest", "plugin", "html", "report", "charts", "enhanced", "testing", "qa", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pytest>=7.0.0",
    "pytest-html>=4.0.0",
    "pytest-metadata>=3.0.0",
    "colorama>=0.4.0",
    "prettytable>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/nireshs/pytest-dashboard"
"Bug Tracker" = "https://github.com/nireshs/pytest-dashboard/issues"
Documentation = "https://github.com/nireshs/pytest-dashboard#readme"
"Source Code" = "https://github.com/nireshs/pytest-dashboard"

[project.entry-points.pytest11]
dashboard = "pytest_dashboard.plugin"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
